src/auth/requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-MySQLdb==1.0.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql==0.0.3
mysqlclient==2.1.1
PyJWT==2.7.0
Werkzeug==2.3.6
zipp==3.15.0
